Add Phone Home capability and refactor code
*Description:*
Added a new job to send usage data to Black Duck. This job runs on the
same schedule as the push scan data task. In the process refactored code
so that PropertyConstants and AttributeConstants were no longer
statically accessible (accessing the static variables when the spring
bean was being instantiated caused null return values and consequently
validation method were failing).

* Edited application.properties and application_prod.properties to have
a placeholder for the plugin version, and edited the build.gradle to
fill in that placeholder using values from the project.properties.
* Ensure PropertyConstants and AttributeConstants are loaded as Spring
Beans
* Removed `static` keyword from all PropertyConstants and
AttributeConstants fields and methods.
* Edited all locations where PropertyConstants and/or AttributeConstants
were used to either provide a Dependency Injected bean or pass the bean
through the constructor for non- spring bean classes
* Added new Job Configuration and Step classes for Phone Home
capability. Job runs on the same schedule as the push scan data task
* Modified all JUnit tests to use Dependency Injected beans where
possible